Move VecDeque Drain iterator to new file
